UVC-010: Code format and cleanup
