+++ umlet-vscode/.gitignore
@@ -1,5 +1,4 @@
 node_modules/
 node/
 npm/
-*.vsix
-out/
+dist/
