--- umlet-vscode/helloworld.txt
@@ -1,1 +1,0 @@
-Hello Worldll!
