+++ umlet-vscode/package.json
@@ -1,96 +1,97 @@
 {
-	"name": "umlet",
-	"displayName": "UMLet",
-	"description": "Free UML Tool for Fast UML Diagrams",
-	"version": "0.3.1",
-	"publisher": "theumletteam",
-	"author": {
-		"name": "The UMLet Team"
-	},
-	"icon": "icons/UMLet_logo_128x128.png",
-	"extensionKind": [
-		"ui"
-	],
-	"engines": {
-		"vscode": "^1.44.0"
-	},
-	"categories": [
-		"Other"
-	],
-	"activationEvents": [
-		"onCustomEditor:uxfCustoms.umletEditor",
-		"onCommand:umlet.createNewDiagram",
-		"onStartupFinished"
-	],
-	"main": "./out/extension.js",
-	"contributes": {
-		"commands": [
-			{
-				"command": "umlet.exportPngOne",
-				"title": "UMLet: Export As PNG.."
-			},
-			{
-				"command": "umlet.exportPngFour",
-				"title": "UMLet: Export As PNG (Scale x4).."
-			},
-			{
-				"command": "umlet.createNewDiagram",
-				"title": "UMLet: Create new diagram"
-			}
-		],
-		"menus": {
-			"commandPalette": [
-				{
-					"command": "umlet.exportPngOne"
-				},
-				{
-					"command": "umlet.exportPngFour"
-				}
-			]
-		},
-		"customEditors": [
-			{
-				"viewType": "uxfCustoms.umletEditor",
-				"displayName": "UMLet",
-				"selector": [
-					{
-						"filenamePattern": "*.uxf"
-					}
-				],
-				"priority": "default"
-			}
-		]
-	},
-	"scripts": {
-		"vsce": "./node_modules/.bin/vsce package -o target/",
-		"vscode:prepublish": "npm run compile",
-		"compile": "tsc -p ./",
-		"lint": "eslint src --ext ts",
-		"watch": "tsc -watch -p ./",
-		"pretest": "npm run compile && npm run lint",
-		"test": "node ./out/test/runTest.js"
-	},
-	"devDependencies": {
-		"@types/file-saver": "^2.0.1",
-		"@types/glob": "^7.1.1",
-		"@types/mocha": "^7.0.1",
-		"@types/node": "^12.12.32",
-		"@types/vscode": "^1.44.0",
-		"@typescript-eslint/eslint-plugin": "^2.25.0",
-		"@typescript-eslint/parser": "^2.25.0",
-		"electron": "^8.2.0",
-		"eslint": "^6.8.0",
-		"glob": "^7.1.6",
-		"mocha": "^7.1.1",
-		"typescript": "^3.7.5",
-		"vscode-test": "^1.3.0"
-	},
-	"dependencies": {
-		"atob": "^2.1.2",
-		"btoa": "^1.2.1",
-		"btob": "^0.1.1",
-		"file-saver": "^2.0.2",
-		"sync-pom-version-to-package": "^1.5.0",
-		"vsce": "^1.75.0"
-	}
+  "name": "umlet",
+  "displayName": "UMLet",
+  "description": "Free UML Tool for Fast UML Diagrams",
+  "version": "0.3.1",
+  "publisher": "theumletteam",
+  "author": {
+    "name": "The UMLet Team"
+  },
+  "icon": "icons/UMLet_logo_128x128.png",
+  "extensionKind": [
+    "ui"
+  ],
+  "engines": {
+    "vscode": "^1.44.0"
+  },
+  "categories": [
+    "Other"
+  ],
+  "activationEvents": [
+    "onCustomEditor:uxfCustoms.umletEditor",
+    "onCommand:umlet.createNewDiagram",
+    "onStartupFinished"
+  ],
+  "main": "./target/out/extension.js",
+  "contributes": {
+    "commands": [
+      {
+        "command": "umlet.exportPngOne",
+        "title": "UMLet: Export As PNG.."
+      },
+      {
+        "command": "umlet.exportPngFour",
+        "title": "UMLet: Export As PNG (Scale x4).."
+      },
+      {
+        "command": "umlet.createNewDiagram",
+        "title": "UMLet: Create new diagram"
+      }
+    ],
+    "menus": {
+      "commandPalette": [
+        {
+          "command": "umlet.exportPngOne"
+        },
+        {
+          "command": "umlet.exportPngFour"
+        }
+      ]
+    },
+    "customEditors": [
+      {
+        "viewType": "uxfCustoms.umletEditor",
+        "displayName": "UMLet",
+        "selector": [
+          {
+            "filenamePattern": "*.uxf"
+          }
+        ],
+        "priority": "default"
+      }
+    ]
+  },
+  "scripts": {
+    "vsce": "./node_modules/.bin/vsce package -o target/",
+    "vscode:prepublish": "npm run compile",
+    "compile": "tsc -p ./",
+    "lint": "eslint src --ext ts",
+    "watch": "tsc -watch -p ./",
+    "pretest": "npm run compile && npm run lint",
+    "test": "node ./out/test/runTest.js"
+  },
+  "devDependencies": {
+    "@types/file-saver": "^2.0.1",
+    "@types/glob": "^7.1.1",
+    "@types/mocha": "^7.0.1",
+    "@types/node": "^12.12.32",
+    "@types/vscode": "^1.44.0",
+    "@typescript-eslint/eslint-plugin": "^2.25.0",
+    "@typescript-eslint/parser": "^2.25.0",
+    "electron": "^8.2.0",
+    "eslint": "^6.8.0",
+    "glob": "^7.1.6",
+    "mocha": "^7.1.1",
+    "typescript": "^3.7.5",
+    "vscode-test": "^1.3.0"
+  },
+  "dependencies": {
+    "atob": "^2.1.2",
+    "btoa": "^1.2.1",
+    "btob": "^0.1.1",
+    "fast-xml-parser": "^3.17.4",
+    "file-saver": "^2.0.2",
+    "sync-pom-version-to-package": "^1.5.0",
+    "vsce": "^1.75.0"
+  }
 }
