+++ umlet-vscode/pom.xml
@@ -33,7 +33,7 @@
 
     <build>
         <plugins>
-            <!-- assemble UMLetino output as zip file -->
+            <!-- assemble UMLetVSCode output as zip file -->
             <plugin>
                 <artifactId>maven-assembly-plugin</artifactId>
                 <version>2.5.3</version>
@@ -72,7 +72,7 @@
                         <configuration>
                             <style>PRETTY</style>
                             <tasks>
-                                <copy file="${project.basedir}/src/main/webapp/umletino.css"
+                                <copy file="${project.basedir}/src/main/web/umletino.css"
                                       todir="${project.build.directory}/${project.build.finalName}"/>
                             </tasks>
                         </configuration>
@@ -116,60 +116,4 @@
             </plugin>
         </plugins>
     </build>
-
-    <!-- this profile can be used to start the devmode for debugging the GWT
-        application -->
-    <profiles>
-        <profile>
-            <id>gwtdevmode</id>
-            <build>
-                <plugins>
-                    <!-- copy static web resources (perhaps there is a better way, but I
-                        haven't found one with the gwt-maven-plugin) -->
-                    <plugin>
-                        <artifactId>maven-antrun-plugin</artifactId>
-                        <executions>
-                            <execution>
-                                <phase>initialize</phase>
-                                <configuration>
-                                    <style>PRETTY</style>
-                                    <tasks>
-                                        <mkdir dir="${project.basedir}/src/main/webapp" />
-                                        <copy todir="${project.build.directory}/gwt/devmode/war">
-                                            <fileset dir="${project.basedir}/src/main/webapp" />
-                                        </copy>
-                                    </tasks>
-                                </configuration>
-                                <goals>
-                                    <goal>run</goal>
-                                </goals>
-                            </execution>
-                        </executions>
-                    </plugin>
-                    <!-- configure devmode -->
-                    <plugin>
-                        <groupId>net.ltgt.gwt.maven</groupId>
-                        <artifactId>gwt-maven-plugin</artifactId>
-                        <version>${gwt.plugin.version}</version>
-                        <extensions>true</extensions>
-                        <executions>
-                            <execution>
-                                <phase>initialize</phase>
-                                <goals>
-                                    <goal>devmode</goal>
-                                </goals>
-                                <configuration>
-                                    <style>PRETTY</style>
-                                    <moduleName>${gwt.module.name}</moduleName>
-                                    <startupUrls>
-                                        <startupUrl>umletino.html</startupUrl>
-                                    </startupUrls>
-                                </configuration>
-                            </execution>
-                        </executions>
-                    </plugin>
-                </plugins>
-            </build>
-        </profile>
-    </profiles>
 </project>
