+++ umlet-vscode/src/main/java/com/vscode/gwt/client/logging/VsCodeLogger.java
@@ -9,63 +9,63 @@ import java.util.logging.Logger;
 
 public class VsCodeLogger implements CustomLogger {
 
-    private int levelValue;
+	private int levelValue;
 
-    @Override
-    public void trace(String message) {
-        if (levelValue == Level.ALL.intValue() || levelValue <= Level.FINEST.intValue()) {
-            postLog(generatePrefixMessage() + "trace|" + message);
-        }
-    }
+	@Override
+	public void trace(String message) {
+		if (levelValue == Level.ALL.intValue() || levelValue <= Level.FINEST.intValue()) {
+			postLog(generatePrefixMessage() + "trace|" + message);
+		}
+	}
 
-    @Override
-    public void debug(String message) {
-        if (levelValue == Level.ALL.intValue() || levelValue <= Level.FINE.intValue()) {
-            postLog(generatePrefixMessage() + "debug|" + message);
-        }
-    }
+	@Override
+	public void debug(String message) {
+		if (levelValue == Level.ALL.intValue() || levelValue <= Level.FINE.intValue()) {
+			postLog(generatePrefixMessage() + "debug|" + message);
+		}
+	}
 
-    @Override
-    public void debug(String message, Throwable throwable) {
-        // Ignoring throwable
-        debug(message);
-    }
+	@Override
+	public void debug(String message, Throwable throwable) {
+		// Ignoring throwable
+		debug(message);
+	}
 
-    @Override
-    public void info(String message) {
-        if (levelValue == Level.ALL.intValue() || levelValue <= Level.INFO.intValue()) {
-            postLog(generatePrefixMessage() + "info|" + message);
-        }
-    }
+	@Override
+	public void info(String message) {
+		if (levelValue == Level.ALL.intValue() || levelValue <= Level.INFO.intValue()) {
+			postLog(generatePrefixMessage() + "info|" + message);
+		}
+	}
 
-    @Override
-    public void error(String message) {
-        if (levelValue == Level.ALL.intValue() || levelValue <= Level.SEVERE.intValue()) {
-            postLog(generatePrefixMessage() + "error|" + message);
-        }
-    }
+	@Override
+	public void error(String message) {
+		if (levelValue == Level.ALL.intValue() || levelValue <= Level.SEVERE.intValue()) {
+			postLog(generatePrefixMessage() + "error|" + message);
+		}
+	}
 
-    @Override
-    public void error(String message, Throwable throwable) {
-        // Ignoring throwable
-        error(message);
-    }
+	@Override
+	public void error(String message, Throwable throwable) {
+		// Ignoring throwable
+		error(message);
+	}
 
-    @Override
-    public void init(Class<?> clazz) {
-        this.levelValue = Logger.getLogger("").getLevel().intValue();
-    }
+	@Override
+	public void init(Class<?> clazz) {
+		this.levelValue = Logger.getLogger("").getLevel().intValue();
+	}
 
-    private String generatePrefixMessage() {
-        Date date = new Date();
-        return "UMLet|" + DateTimeFormat.getFormat("yyyy-MM-dd HH:mm:ss").format(date) + "|";
-    }
+	private String generatePrefixMessage() {
+		Date date = new Date();
+		return "UMLet|" + DateTimeFormat.getFormat("yyyy-MM-dd HH:mm:ss").format(date) + "|";
+	}
 
-    private native void postLog(String message) /*-{
-        window.parent.vscode.postMessage({
-            command: 'postLog',
-            text: message
-        });
-    }-*/;
+	private native void postLog(String message) /*-{
+		window.parent.vscode.postMessage({
+			command: 'postLog',
+			text: message
+		});
+	}-*/;
 
 }
