+++ umlet-vscode/src/main/java/com/vscode/gwt/client/resources/VsCodeHelptextResources.java
@@ -4,6 +4,6 @@ import com.baselet.gwt.client.resources.HelptextResources;
 import com.google.gwt.resources.client.TextResource;
 
 public interface VsCodeHelptextResources extends HelptextResources {
-    @Source("Helptext.txt")
-    TextResource helpText();
+	@Source("Helptext.txt")
+	TextResource helpText();
 }
