+++ umlet-vscode/src/main/java/com/vscode/gwt/client/view/VsCodeMainView.java
@@ -3,8 +3,8 @@ package com.vscode.gwt.client.view;
 import com.baselet.gwt.client.view.MainView;
 
 public class VsCodeMainView extends MainView {
-    public VsCodeMainView() {
-        super();
-        diagramPaletteSplitter.setWidgetHidden(diagramPaletteSplitter.getWidget(0), true);
-    }
+	public VsCodeMainView() {
+		super();
+		diagramPaletteSplitter.setWidgetHidden(diagramPaletteSplitter.getWidget(0), true);
+	}
 }
