--- umlet-vscode/src/main/webapp/index.html
@@ -1,125 +1,0 @@
-
-<html>
-<head>
-<title>
-UMLetino - Free Online UML Tool for Fast UML Diagrams
-</title>
-
-<!-- Place this tag in your head or just before your close body tag. -->
-<script src="https://apis.google.com/js/platform.js" async defer></script>
-</head>
-
-<body>
-
-<table border="0" align="center" width="728">
-
-
-
-<tr>
-<th colspan="2">
-    <script type="text/javascript"><!--
-		google_ad_client = "pub-7477443223564915";
-		google_ad_width = 728;
-		google_ad_height = 90;
-		google_ad_format = "728x90_as";
-		google_ad_type = "text_image";
-		google_ad_channel = "";
-		//-->
-	</script>
-	<script type="text/javascript"
-	  src="http://pagead2.googlesyndication.com/pagead/show_ads.js">
-	</script>
-</th>
-</tr>
-
-
-<tr>
-<th colspan="2">
-    <br/>
-    <br/>
-    <br/>
-</th>
-</tr>
-
-
-<tr>
-<td>
-    <font face="Verdana, Arial, Helvetica, sans-serif">
-    <p>
-    
-	<b>UMLetino</b> is a free online UML tool for fast UML diagrams. It runs in your browser, and does not require any installs.</br>
-    It is based on <a href="http://www.umlet.com">UMLet</a> (which is available as stand-alone tool or Eclipse plugin), and shares its fast, text-based way of drawing UML sketches.</br>
-    Main features: install-free web app; save diagrams in browser storage; support for many UML diagram types; simple, markup-based UML element modifications; png export.</br>
-    </br>
-    <i>New in current version 14.2: improved relations, bugfixes</i></br>
-    </br>
-    <a href="umletino.html" target="_blank">Start UMLetino NOW!</a> (JavaScript and Cookies required)
-    
-    </p>
-    </font>
-</td>
-<td>
-    <p align="center">
-    <a href="http://www.umletino.com"><img src="UMLetino_logo_small.png" border="0"></a>
-    </p>
-</td>
-</tr>
-
-
-<tr>
-<th colspan="2" align="left">
-
-<p>
-<iframe src="http://www.facebook.com/plugins/like.php?href=http%3A%2F%2Fwww.umletino.com&amp;layout=button_count&amp;action=like&amp;show_faces=false&amp;share=false&amp;height=21" scrolling="no" frameborder="0" style="border:none; overflow:hidden; height:21px;" allowTransparency="true"></iframe>
-</p>
-
-<p>
-<a href="https://twitter.com/twumlet" class="twitter-follow-button" data-show-count="false">Follow @twumlet</a>
-<script>!function(d,s,id){var js,fjs=d.getElementsByTagName(s)[0],p=/^http:/.test(d.location)?'http':'https';if(!d.getElementById(id)){js=d.createElement(s);js.id=id;js.src=p+'://platform.twitter.com/widgets.js';fjs.parentNode.insertBefore(js,fjs);}}(document, 'script', 'twitter-wjs');</script>
-
-<a href="https://twitter.com/share" class="twitter-share-button" data-url="http://www.umletino.com" data-via="twumlet">Tweet</a>
-<script>!function(d,s,id){var js,fjs=d.getElementsByTagName(s)[0],p=/^http:/.test(d.location)?'http':'https';if(!d.getElementById(id)){js=d.createElement(s);js.id=id;js.src=p+'://platform.twitter.com/widgets.js';fjs.parentNode.insertBefore(js,fjs);}}(document, 'script', 'twitter-wjs');</script>
-</p>
-
-<p>
-<!-- Place this tag where you want the +1 button to render. -->
-<div class="g-plusone" data-size="medium" data-href="http://www.umletino.com"></div>
-</p>
-
-</th>
-</tr>
-
-
-<tr>
-<th colspan="2">
-    <br/>
-    <br/>
-    <br/>
-</th>
-</tr>
-
-
-<tr>
-<th colspan="2">
-    <script type="text/javascript"><!--
-		google_ad_client = "pub-7477443223564915";
-		google_ad_width = 728;
-		google_ad_height = 90;
-		google_ad_format = "728x90_as";
-		google_ad_type = "text_image";
-		google_ad_channel = "";
-		//-->
-	</script>
-	<script type="text/javascript"
-	  src="http://pagead2.googlesyndication.com/pagead/show_ads.js">
-	</script>
-</th>
-</tr>
-
-
-
-
-</table>
-
-</body>
-</html>
