--- umlet-vscode/src/main/webapp/umletino.html
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-  	<meta name="viewport" content="user-scalable=no" />
-    <meta http-equiv="content-type" content="text/html; charset=UTF-8">
-    <link type="text/css" rel="stylesheet" href="umletino.css">
-    <link rel="icon" type="image/x-icon" href="favicon.ico">
-    <title>UMLetino - Free Online UML Tool for Fast UML Diagrams</title>
-    <script type="text/javascript" src="umletvscode/umletvscode.nocache.js?${build.time.urlsafe}"></script>
-  </head>
-  <body>
-    <!-- the following line is necessary for history support -->
-    <iframe src="javascript:''" id="__gwt_historyFrame" tabIndex='-1' style="position:absolute;width:0;height:0;border:0"></iframe>
-    
-    <!-- the website will not work without JavaScript -->
-    <noscript>
-    	<div style="width: 25em; position: absolute; left: 50%; margin-left: -11em; background-color: white; border: 1px solid red; padding: 4px; font-family: sans-serif">
-	    	You must enable JavaScript to use this web application.
-		</div>
-    </noscript>
-    <div align="left" id="featurewarning" style="color: red; font-family: sans-serif; font-weight:bold; font-size:1.2em"></div>
-  </body>
-</html>
