+++ umlet-vscode/tsconfig.json
@@ -2,7 +2,7 @@
 	"compilerOptions": {
 		"module": "commonjs",
 		"target": "es6",
-		"outDir": "out",
+		"outDir": "target/out",
 		"lib": [
 			"es6",
             "dom"
